1. Fixed issue where covariance of sampled FIRM node was not being saved in graph.
2. Added Graph to XML function in FIRM utils and corresponding in FIRM planner.

+++ include/Weight/FIRMWeight.h
@@ -93,7 +93,7 @@ class FIRMWeight {
   protected:
     double  cost_; // the cost of traversing the edge
 
-    size_t controllerID_; // the identifier of the controller associated with the edge
+    int controllerID_; // the identifier of the controller associated with the edge
 
     double successProbability_; //  the transition probability of the edge
